fix: Add reusable hook for screen-size and useState
