+++ backend/src/middleware/cors.middleware.ts
@@ -9,7 +9,7 @@ import { cors } from 'hono/cors';
 
 // configure cors middleware
 const corsMiddleware = cors({
-  origin: ['http://localhost:5173', 'http://127.0.0.1:5173'], // you need to change this to the ip you are running the frontend with (for example i was using 5173. but for you, you need to change it to 5174)
+  origin: ['http://localhost:5174', 'http://127.0.0.1:5173'], // you need to change this to the ip you are running the frontend with (for example i was using 5173. but for you, you need to change it to 5174)
   allowHeaders: ['Content-Type', 'Authorization'],
   allowMethods: ['GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'],
   credentials: true,
